third_party/blink/web_tests/html/sections/aside-element-expected.txt

Various tests for the aside element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

<aside> closes <p>:
PASS aside1.parentNode.nodeName == "p" is false
<p> does not close <aside>:
PASS p1.parentNode.nodeName is "ASIDE"
<aside> can be nested inside <aside>:
PASS aside3.parentNode.id is "aside2"
Residual style:
PASS getWeight("aside4") is "700"
PASS getWeight("span1") is "700"
FormatBlock:
PASS document.getElementById("span2").parentNode.nodeName is "ASIDE"
PASS successfullyParsed is true

TEST COMPLETE
